deb 11.03

+++ headers/cycle.h
@@ -10,7 +10,7 @@ typedef struct
 {
 	WORD last_ptr_rx_buf[4];
 	BYTE stage[4];
-	WORD time[4];
+	DWORD time[4];
 }VAR;
 #pragma pack()
